requirements.txt
Django==1.7
Whoosh==2.6.0
django-autoslug==1.7.2
django-haystack==2.3.1
